+++ inst/NEWS.md
@@ -1,8 +1,16 @@
+# rdecision 1.2.1.9007
+
+* Decision tree drawing algorithm improved to place leaf node labels to the
+  right of the symbol, and to accurately calculate margin sizes required to
+  accommodate symbols and labels efficiently.
+* Added package qualifier for function dev.size in grDevices, and functions
+  abline, axis, rect, strwidth and text in package graphics.
+
 # rdecision 1.2.1.9006
 
 * Moved code to plot a tornado diagram into 'utils.R` to make it available
   for creating such diagrams given any input source (e.g., Markov models),
-  rather than restricing its use to `ModVar`s in decision trees.
+  rather than restricting its use to `ModVar`s in decision trees.
 
 # rdecision 1.2.1.9005